Add support for Y10B pixelformat
Add support for Y10B, a 10 bit per pixel greyscale format, the stream is
represented using a bit-packed array with each element of 10bit.

This format is used by the Kinect Sensor device and its gspca driver
uses it for the Infra Red stream.

We unpack the array to elements of 2 bytes before converting to YUYV.

Signed-off-by: Antonio Ospite <ospite@studenti.unina.it>